chat/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
websockets==14.1
pyjwt==2.10.1
python-dotenv==1.0.1
redis==5.2.1
sqlmodel==0.0.14
asyncpg==0.29.0
tzdata
